utils: Move Git index plumbing helpers
The Git utility layer has named modules for object IO and repository
identity, while index and tree plumbing still share the command execution
wrapper.

The project lacks a named boundary for index mutations and temporary index
tree construction. Batch, undo, and staging code must import those
operations from the broad command module.

This commit addresses that boundary gap by moving GitIndexEntryUpdate,
temporary index setup, index updates, tree writing, commit-tree creation,
apply-to-index, add-path staging, and path reset helpers into
utils.git_index, then updating source callers.

Subsequent commits will move index-focused tests and add an architecture
guard for the Git index helper boundary.
