+++ src/git_stage_batch/batch/state_refs.py
@@ -13,13 +13,15 @@ from ..core.buffer import LineBuffer
 from ..utils.repository_buffers import load_git_object_as_buffer
 from ..utils.file_io import read_text_file_contents
 from ..utils.git import (
+    run_git_command,
+    update_git_refs,
+)
+from ..utils.git_index import (
     GitIndexEntryUpdate,
     git_commit_tree,
     git_update_index_entries,
     git_write_tree,
-    run_git_command,
     temp_git_index,
-    update_git_refs,
 )
 from ..utils.git_object_io import create_git_blob
 from ..utils.paths import get_batch_metadata_file_path
