+++ src/git_stage_batch/commands/include_from.py
@@ -42,7 +42,7 @@ from ..exceptions import (
     exit_with_error,
 )
 from ..i18n import _
-from ..utils.git import git_refresh_index
+from ..utils.git_index import git_refresh_index
 from ..utils.git_repository import require_git_repository
