+++ src/git_stage_batch/commands/unblock_file.py
@@ -19,10 +19,8 @@ from ..data.ignore_files import (
 from ..exceptions import NoMoreHunks, exit_with_error
 from ..i18n import _
 from ..utils.file_io import append_file_path_to_file, read_file_paths_file, remove_file_path_from_file
-from ..utils.git import (
-    git_add_paths,
-    run_git_command,
-)
+from ..utils.git import run_git_command
+from ..utils.git_index import git_add_paths
 from ..utils.git_repository import (
     require_git_repository,
     resolve_file_path_to_repo_relative,
