+++ src/git_stage_batch/data/batch_sources.py
@@ -19,13 +19,15 @@ from ..exceptions import CommandError
 from ..i18n import _
 from ..utils.file_io import read_file_paths_file, read_text_file_contents, write_text_file_contents
 from ..utils.git import (
+    run_git_command,
+    update_git_refs,
+)
+from ..utils.git_index import (
     git_commit_tree,
     git_read_tree,
     git_update_index,
     git_write_tree,
-    run_git_command,
     temp_git_index,
-    update_git_refs,
 )
 from ..utils.git_repository import get_git_repository_root_path
 from ..utils.git_object_io import create_git_blob, list_git_tree_blobs
