+++ src/git_stage_batch/data/session.py
@@ -16,10 +16,10 @@ from ..utils.file_io import (
     write_text_file_contents,
 )
 from ..utils.git import (
-    git_add_paths,
     git_remove_paths,
     run_git_command,
 )
+from ..utils.git_index import git_add_paths
 from ..utils.git_repository import get_git_repository_root_path
 from ..utils.journal import log_journal
 from ..utils.paths import (
